pending purchase

+++ app/Http/Controllers/PurchaseController.php
@@ -28,10 +28,11 @@ class PurchaseController extends Controller
         $suppliers = Supplier::all();
         $companies = Company::all();
         $mod = new Purchase();
-        if($user->role->slug == 'user'){
+        if($user->hasRole('user') || $user->hasRole('secretary')){
             $mod = $user->company->purchases();
             $stores = $user->company->stores;
         }
+        $mod = $mod->where('status', 1);
         $company_id = $reference_no = $supplier_id = $store_id = $period = $expiry_period = $keyword = '';
         $sort_by_date = 'desc';
         if ($request->get('company_id') != ""){
@@ -80,7 +81,6 @@ class PurchaseController extends Controller
         if($request->sort_by_date != ''){
             $sort_by_date = $request->sort_by_date;
         }
-        // dump($sort_by_date);
         $pagesize = session('pagesize');
         $data = $mod->orderBy('timestamp', $sort_by_date)->paginate($pagesize);
         return view('purchase.index', compact('data', 'companies', 'stores', 'suppliers', 'company_id', 'store_id', 'supplier_id', 'reference_no', 'period', 'expiry_period', 'keyword', 'sort_by_date'));
@@ -88,11 +88,15 @@ class PurchaseController extends Controller
 
     public function create(Request $request){
         config(['site.page' => 'purchase_create']);
-        $user = Auth::user();  
+        $user = Auth::user();
+        $allowed = ['user', 'secretary'];
+        if(!in_array($user->role->slug, $allowed)){
+            return back()->withErrors(['role_error' => __('page.not_allowed_page')]);
+        }
         $suppliers = Supplier::all();
         $products = Product::all();
         $stores = Store::all();
-        if($user->hasRole('user')){
+        if($user->hasRole('user') || $user->hasRole('secretary')){
             $stores = $user->company->stores;
         }
         return view('purchase.create', compact('suppliers', 'stores', 'products'));
@@ -126,7 +130,7 @@ class PurchaseController extends Controller
             $item->expiry_date = date('Y-m-d', strtotime("+".$data['credit_days']."days", strtotime($item->timestamp)));
         }        
         $item->credit_days = $data['credit_days'];
-        $item->status = $data['status'];
+        // $item->status = $data['status'];
         $item->note = $data['note'];
 
         if($request->has("attachment")){
@@ -179,12 +183,16 @@ class PurchaseController extends Controller
 
     public function edit(Request $request, $id){    
         config(['site.page' => 'purchase']); 
-        $user = Auth::user();   
+        $user = Auth::user(); 
+        $allowed = ['user', 'admin'];
+        if(!in_array($user->role->slug, $allowed)){
+            return back()->withErrors(['role_error' => __('page.not_allowed_page')]);
+        }  
         $purchase = Purchase::find($id);        
         $suppliers = Supplier::all();
         $products = Product::all();
         $stores = Store::all();
-        if($user->role->slug == 'user'){
+        if($user->hasRole('user')){
             $stores = $user->company->stores;
         }
 
@@ -267,10 +275,92 @@ class PurchaseController extends Controller
     }
 
     public function delete($id){
+        $user = Auth::user();
+        $allowed = ['user', 'admin'];
+        if(!in_array($user->role->slug, $allowed)){
+            return back()->withErrors(['role_error' => __('page.not_allowed_page')]);
+        }
         $item = Purchase::find($id);
         $item->orders()->delete();
         $item->payments()->delete();
         $item->delete();
         return back()->with("success", __('page.deleted_successfully'));
+    }
+
+    public function pending_purchases(Request $request){
+        config(['site.page' => 'pending_purchases']);
+        $user = Auth::user();
+        $stores = Store::all();
+        $suppliers = Supplier::all();
+        $companies = Company::all();
+        $mod = new Purchase();
+        if($user->hasRole('user') || $user->hasRole('secretary')){
+            $mod = $user->company->purchases();
+            $stores = $user->company->stores;
+        }
+        $mod = $mod->where('status', 0);
+        $company_id = $reference_no = $supplier_id = $store_id = $period = $expiry_period = $keyword = '';
+        $sort_by_date = 'desc';
+        if ($request->get('company_id') != ""){
+            $company_id = $request->get('company_id');
+            $mod = $mod->where('company_id', $company_id);
+        }
+        if ($request->get('reference_no') != ""){
+            $reference_no = $request->get('reference_no');
+            $mod = $mod->where('reference_no', 'LIKE', "%$reference_no%");
+        }
+        if ($request->get('supplier_id') != ""){
+            $supplier_id = $request->get('supplier_id');
+            $mod = $mod->where('supplier_id', $supplier_id);
+        }
+        if ($request->get('store_id') != ""){
+            $store_id = $request->get('store_id');
+            $mod = $mod->where('store_id', $store_id);
+        }
+        if ($request->get('period') != ""){   
+            $period = $request->get('period');
+            $from = substr($period, 0, 10);
+            $to = substr($period, 14, 10);
+            $mod = $mod->whereBetween('timestamp', [$from, $to]);
+        }
+        if ($request->get('expiry_period') != ""){   
+            $expiry_period = $request->get('expiry_period');
+            $from = substr($expiry_period, 0, 10);
+            $to = substr($expiry_period, 14, 10);
+            $mod = $mod->whereBetween('expiry_date', [$from, $to]);
+        }
+        if ($request->get('keyword') != ""){
+            $keyword = $request->keyword;
+            $company_array = Company::where('name', 'LIKE', "%$keyword%")->pluck('id');
+            $supplier_array = Supplier::where('company', 'LIKE', "%$keyword%")->pluck('id');
+            $store_array = Store::where('name', 'LIKE', "%$keyword%")->pluck('id');
+
+            $mod = $mod->where(function($query) use($keyword, $company_array, $store_array, $supplier_array){
+                return $query->where('reference_no', 'LIKE', "%$keyword%")
+                        ->orWhereIn('company_id', $company_array)
+                        ->orWhereIn('store_id', $store_array)
+                        ->orWhereIn('supplier_id', $supplier_array)
+                        ->orWhere('timestamp', 'LIKE', "%$keyword%")
+                        ->orWhere('grand_total', 'LIKE', "%$keyword%");
+            });
+        }
+        if($request->sort_by_date != ''){
+            $sort_by_date = $request->sort_by_date;
+        }
+        $pagesize = session('pagesize');
+        $data = $mod->orderBy('timestamp', $sort_by_date)->paginate($pagesize);
+        return view('purchase.pending', compact('data', 'companies', 'stores', 'suppliers', 'company_id', 'store_id', 'supplier_id', 'reference_no', 'period', 'expiry_period', 'keyword', 'sort_by_date'));
+    }
+
+    public function approve($id){
+        $user = Auth::user();
+        $allowed = ['user', 'admin'];
+        if(!in_array($user->role->slug, $allowed)){
+            return back()->withErrors(['role_error' => __('page.not_allowed_page')]);
+        }
+
+        $item = Purchase::find($id);
+        $item->update(['status' => 1]);
+        return back()->with("success", __('page.approved_successfully'));
     }
 }
